requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
yt-dlp==2023.10.13
moviepy>=1.0.3
celery==5.3.1
redis==5.0.0
django-celery-results==2.5.1
django-celery-beat==2.5.0
Pillow>=10.2.0
requests==2.31.0
psycopg2-binary>=2.9.9